feat(types): 型定義の分散と一貫性の欠如を解決
- src/types/forms.ts を新規作成してフォーム関連型を集約
- src/types/slack.ts を整理し、Hook関連型を追加
- src/types/error.ts の構造を改善し、コメントを充実
- 各コンポーネントから重複型定義を削除し、統一的な型管理を実現
- TypeScript コンパイルエラーなし、Lint チェック通過
