+++ src/components/SlackSearchForm.tsx
@@ -13,45 +13,7 @@ import { SlackResultsDisplay } from '@/components/SlackResultsDisplay'
 import { SlackSearchActions } from '@/components/SlackSearchActions'
 import { SlackSearchInput } from '@/components/SlackSearchInput'
 import { SlackTokenInput } from '@/components/SlackTokenInput'
-import type { ProgressStatus } from '@/hooks/useSlackSearchUnified'
-import type { SlackThread } from '@/types/slack'
-
-type SlackSearchFormProps = {
-  // 検索条件
-  searchQuery: string
-  onSearchQueryChange: (query: string) => void
-  token: string
-  onTokenChange: (token: string) => void
-
-  // 詳細フィルター
-  showAdvanced: boolean
-  onToggleAdvanced: () => void
-  channel: string
-  onChannelChange: (channel: string) => void
-  author: string
-  onAuthorChange: (author: string) => void
-  startDate: string
-  onStartDateChange: (date: string) => void
-  endDate: string
-  onEndDateChange: (date: string) => void
-
-  // 状態
-  isLoading: boolean
-  isDownloading: boolean
-  progressStatus: ProgressStatus
-  hasSearched: boolean
-  error: string | null
-
-  // 結果
-  slackThreads: SlackThread[]
-  userMaps: Record<string, string>
-  permalinkMaps: Record<string, string>
-
-  // イベントハンドラー
-  onSubmit: (e: React.FormEvent<HTMLFormElement>) => void
-  onDownload: (markdownContent: string, searchQuery: string, hasContent: boolean) => void
-  onFullDownload: (markdownContent: string, searchQuery: string, hasContent: boolean) => void
-}
+import type { SlackSearchFormProps } from '@/types/forms'
 
 export function SlackSearchForm({
   searchQuery,
